// src/modules/reports/Dto/reports.dto.ts
export class ReportsDto {
    autor_usuario: string;
    autor_cliente: string;
    codigo_acao: string;
    categoria: string;
    operador: string;
    cancelar: boolean;
    cadastrar: boolean;
    editar: boolean;
    login: boolean;
    logout: boolean;
    agendamento: boolean;
    fila: boolean;
    walkin: boolean;
    atendimento: boolean;
    id_cliente: string;
    observacao: string;
}